move chunks into utils
changed limits for max cmd line: 8190 for windows, 32760 otherwise. no consistent for both regex_files_rg and compile_missing_pyc
added MAX_CMD_LINE_LENGTH variable to avoid duplication
